add SpendProfileOrganization and move other SpendProfile model classes into dedicated directory
