+++ service/src/main/java/bio/terra/workspace/service/spendprofile/model/SpendProfileOrganization.java
@@ -1,0 +1,9 @@
+package bio.terra.workspace.service.spendprofile.model;
+
+import bio.terra.profile.model.Organization;
+
+public record SpendProfileOrganization(Boolean enterprise) {
+  public SpendProfileOrganization(Organization organization) {
+    this(organization.isEnterprise());
+  }
+}
