+++ service/src/main/java/bio/terra/workspace/service/workspace/exceptions/NoAzureAppCoordinatesException.java
@@ -1,7 +1,7 @@
 package bio.terra.workspace.service.workspace.exceptions;
 
 import bio.terra.common.exception.BadRequestException;
-import bio.terra.workspace.service.spendprofile.SpendProfileId;
+import bio.terra.workspace.service.spendprofile.model.SpendProfileId;
 
 public class NoAzureAppCoordinatesException extends BadRequestException {
   public NoAzureAppCoordinatesException(String message) {
