+++ service/src/main/java/bio/terra/workspace/service/workspace/model/CloudContextCommonFields.java
@@ -4,7 +4,7 @@ import bio.terra.common.exception.ErrorReportException;
 import bio.terra.workspace.app.controller.shared.ControllerUtils;
 import bio.terra.workspace.generated.model.ApiOperationState;
 import bio.terra.workspace.service.resource.model.WsmResourceState;
-import bio.terra.workspace.service.spendprofile.SpendProfileId;
+import bio.terra.workspace.service.spendprofile.model.SpendProfileId;
 
 public record CloudContextCommonFields(
     SpendProfileId spendProfileId,
