+++ service/src/test/java/bio/terra/workspace/service/policy/flight/LinkSpendProfilePolicyAttributesStepTest.java
@@ -23,7 +23,7 @@ import bio.terra.workspace.common.BaseSpringBootUnitTest;
 import bio.terra.workspace.service.policy.TpsApiDispatch;
 import bio.terra.workspace.service.policy.exception.PolicyServiceDuplicateException;
 import bio.terra.workspace.service.resource.exception.PolicyConflictException;
-import bio.terra.workspace.service.spendprofile.SpendProfileId;
+import bio.terra.workspace.service.spendprofile.model.SpendProfileId;
 import bio.terra.workspace.service.workspace.flight.WorkspaceFlightMapKeys;
 import java.util.UUID;
 import org.junit.jupiter.api.Test;
