+++ service/src/test/java/bio/terra/workspace/service/resource/statetests/AwsResourceStateFailureTest.java
@@ -25,7 +25,7 @@ import bio.terra.workspace.generated.model.ApiCreateControlledAwsSageMakerNotebo
 import bio.terra.workspace.generated.model.ApiDeleteControlledAwsResourceRequestBody;
 import bio.terra.workspace.generated.model.ApiJobControl;
 import bio.terra.workspace.service.resource.referenced.ReferencedResourceService;
-import bio.terra.workspace.service.spendprofile.SpendProfileId;
+import bio.terra.workspace.service.spendprofile.model.SpendProfileId;
 import bio.terra.workspace.service.workspace.model.CloudPlatform;
 import bio.terra.workspace.service.workspace.model.Workspace;
 import com.fasterxml.jackson.databind.ObjectMapper;
