+++ service/src/test/java/bio/terra/workspace/service/resource/statetests/AzureResourceStateFailureTest.java
@@ -33,7 +33,7 @@ import bio.terra.workspace.generated.model.ApiCreateControlledAzureVmRequestBody
 import bio.terra.workspace.generated.model.ApiDeleteControlledAzureResourceRequest;
 import bio.terra.workspace.generated.model.ApiJobControl;
 import bio.terra.workspace.service.resource.referenced.ReferencedResourceService;
-import bio.terra.workspace.service.spendprofile.SpendProfileId;
+import bio.terra.workspace.service.spendprofile.model.SpendProfileId;
 import bio.terra.workspace.service.workspace.model.CloudPlatform;
 import bio.terra.workspace.service.workspace.model.Workspace;
 import com.azure.core.management.Region;
